+++ am/src/main.rs
@@ -20,6 +20,7 @@ mod dir;
 mod downloader;
 mod interactive;
 mod server;
+mod terminal;
 
 #[tokio::main]
 async fn main() {